Make cleaner target names + don't get filenames from target names
When .Rmd files have illegal characters (i.e. start with digits, contain "-", etc), the resulting object names don't work in R, so we have to use make.names() to clean those up. But then that makes it so the resulting filenames are based on the cleaned up names, which we don't want, so we have to make them separate processes here.

Thanks to @tjmahr for figuring this out! tjmahr/notestar@f1647ee